requirements.txt
paddlepaddle==2.4.0
paddleocr>=2.0.6
PyPDF2==2.11.1
pandastable==0.13.0
pdf2image==1.16.0
streamlit-option-menu==0.3.2
streamlit-aggrid==0.2.2.post2
